[update] rename to `regression_set` and rewrite
